+++ gcpy/plot/compare_single_level.py
@@ -24,7 +24,8 @@ from gcpy.regrid import regrid_comparison_data, create_regridders
 from gcpy.util import \
     reshape_MAPL_CS, get_diff_of_diffs, get_molwt_from_metadata, \
     all_zero_or_nan, slice_by_lev_and_time, compare_varnames, \
-    read_species_metadata, verify_variable_type
+    read_species_metadata, verify_variable_type, \
+    warn_if_flip_levels_mismatch
 from gcpy.units import check_units, data_unit_is_mol_per_mol
 from gcpy.constants import MW_AIR_g, NO_STRETCH_SG_PARAMS
 from gcpy.plot.core import gcpy_style, six_panel_subplot_names, \
@@ -192,6 +193,8 @@ def compare_single_level(
     # Error check arguments
     verify_variable_type(refdata, xr.Dataset)
     verify_variable_type(devdata, xr.Dataset)
+
+    warn_if_flip_levels_mismatch(flip_ref, flip_dev)
 
     # Create empty lists for keyword arguments
     if extent is None: